refactor: 배럴 파일 제거
- 장점
  - 모듈 import 경로 일관성 유지
  - 깔끔한 import 문 관리

- 단점
  - 전체 모듈이 단일 번들로 묶여 불필요한 코드 포함
  - 동적 임포트 어려움 (필요한 시점에 모듈 로드 제한)
  - 순환 참조 발생 시 디버깅/코드 스플리팅 복잡

현재 프로젝트:
- 번들 최적화(트리 셰이킹, 코드 스플리팅)를 하지 않아 당장은 영향 미미
- 향후 성능 최적화 시 배럴 파일 제거 고려 필요
